requirements.txt
google-genai==1.10.0
notebook==7.4.0
pypdf==5.4.0
python-dotenv==1.1.0
